🧹 Add ci for `fond-of-oryx/customer-product-list-search-rest-api` - add required dependencies to composer.json - add phpstan support - add github actions - add missing unit tests - move plugins to correct folder - ...
