+++ bundles/customer-product-list-search-rest-api/composer.json
@@ -10,11 +10,14 @@
   ],
   "require": {
     "php": ">=8.0",
-    "fond-of-oryx/product-list-search-rest-api-extension": "^3.0.0"
+    "fond-of-oryx/product-list-search-rest-api-extension": "^3.0.0",
+    "fond-of-spryker/product-list-customer": "dev-spryker_upgrade",
+    "spryker/product-list": "^1.0.0"
   },
   "require-dev": {
-    "fond-of-codeception/spryker": "^2.0.0",
-    "spryker/code-sniffer": "^0.12"
+    "fond-of-codeception/spryker": "*",
+    "spryker/code-sniffer": "*",
+    "spryker-sdk/phpstan-spryker": "*"
   },
   "minimum-stability": "dev",
   "prefer-stable": true,
@@ -26,7 +29,8 @@
   "autoload-dev": {
     "psr-4": {
       "FondOfOryx\\": "tests/FondOfOryx/",
-      "Generated\\": "src/Generated/"
+      "Generated\\": "src/Generated/",
+      "Orm\\": "src/Orm/"
     }
   },
   "config": {
