+++ bundles/customer-product-list-search-rest-api/src/FondOfOryx/Shared/CustomerProductListSearchRestApi/Transfer/customer_product_list_search_rest_api.transfer.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0"?>
+<transfers xmlns="spryker:transfer-01"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="spryker:transfer-01 https://static.spryker.com/transfer-01.xsd">
+    <transfer name="RestUser">
+        <property name="surrogateIdentifier" type="int"/>
+    </transfer>
+
+    <transfer name="FilterField">
+        <property name="value" type="string"/>
+        <property name="type" type="string"/>
+    </transfer>
+
+    <transfer name="QueryJoin">
+        <property name="left" type="array" singular="left"/>
+        <property name="right" type="array" singular="right"/>
+        <property name="joinType" type="string"/>
+        <property name="whereConditions" type="QueryWhereCondition[]" singular="queryWhereCondition"/>
+    </transfer>
+
+    <transfer name="QueryJoinCollection">
+        <property name="queryJoins" type="QueryJoin[]" singular="queryJoin"/>
+    </transfer>
+
+    <transfer name="QueryWhereCondition">
+        <property name="column" type="string"/>
+        <property name="value" type="string"/>
+        <property name="comparison" type="string"/>
+    </transfer>
+</transfers>
