+++ bundles/customer-product-list-search-rest-api/src/FondOfOryx/Zed/CustomerProductListSearchRestApi/Communication/Plugin/ProductListSearchRestApiExtension/CustomerSearchProductListQueryExpanderPlugin.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace FondOfOryx\Zed\CustomerProductListSearchRestApi\Communication\Plugin\ProductListSearchRestApi;
+namespace FondOfOryx\Zed\CustomerProductListSearchRestApi\Communication\Plugin\ProductListSearchRestApiExtension;
 
 use FondOfOryx\Shared\CustomerProductListSearchRestApi\CustomerProductListSearchRestApiConstants;
 use FondOfOryx\Zed\ProductListSearchRestApiExtension\Dependency\Plugin\SearchProductListQueryExpanderPluginInterface;
